Using an untagged enum to support u64 for integers
The original code supported type `i64` for integers, where it was not
possible for the type to be resolved to `IntegerType` if maximum value
was `u64::MAX`. Since `i128` do not work well with untagged enums
(`SchemaKind`), using `i128` caused the schema to be resolved to
`AnySchema`, which is not right.

One gotcha is - if the minimum is negative and maximum is `u64::MAX`,
this would resolve the minimum and maximum to two different variants of
the enum, which is not wrong from the de-serialization point of view,
but can be handled by the application using this resolved type.