Add enum field to boolean types
I'm using this library with the stripe API, and they've got a few places
where they use the enum field on a boolean to indicate that a field
will always be either true or false.

This PR updates the types in this crate to support that.  It's a small
change, but still probably counts as a breaking change, so not sure how
you want to handle that?